pyproject.toml
[project]
name = "steelseries-oled"
version = "1.0.0"
description = "Display GIFs and system stats on SteelSeries keyboard OLED screens"
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
authors = [{ name = "DreamTimeZ" }]
keywords = ["steelseries", "oled", "keyboard", "hid", "usb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Hardware",
    "Typing :: Typed",
]
dependencies = [
    "hidapi>=0.15.0",
    "nvidia-ml-py>=13.590.44",
    "pillow>=12.1.0",
    "psutil>=7.0.0",
    "requests>=2.32.5",
]

[project.scripts]
steelseries = "steelseries_oled.cli:main"

[project.urls]
Repository = "https://github.com/DreamTimeZ/steelseries-oled"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/steelseries_oled"]

[dependency-groups]
dev = [
    "mypy>=1.19",
    "pyinstaller>=6.17.0",
    "ruff>=0.14",
    "pre-commit>=4.2.0",
    "types-requests>=2.32",
]
test = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
]

[tool.ruff]
target-version = "py314"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "LOG",    # flake8-logging
    "G",      # flake8-logging-format
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "PL",     # Pylint
    "TRY",    # tryceratops
    "FLY",    # flynt
    "PERF",   # Perflint
    "FURB",   # refurb
    "RUF",    # Ruff-specific
]
ignore = [
    "PLR0911", # Too many return statements (CLI functions need multiple exit paths)
    "PLR0913", # Too many arguments
    "TRY003",  # Avoid specifying long messages outside exception class
    "TRY300",  # Consider moving to else block (not always cleaner)
    "TRY301",  # Abstract raise to inner function (overkill for cleanup patterns)
    "SIM108",  # Ternary operator (readability preference)
    "TC001",   # Move first-party import into TYPE_CHECKING (PEP 649 handles this in 3.14)
    "TC002",   # Move third-party import into TYPE_CHECKING (PEP 649 handles this in 3.14)
    "TC003",   # Move stdlib import into TYPE_CHECKING (PEP 649 handles this in 3.14)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "PLR2004", # Magic value comparisons are fine in tests
    "SIM117",  # Nested with statements are fine in tests
    "ARG",     # Unused arguments for fixtures
    "SLF001",  # Private member access is needed for testing internal state
    "PLC0415", # Function-level imports for lazy loading / type checking
]
"src/steelseries_oled/_signal.py" = ["ARG001"]  # Signal handler requires unused frame arg
"src/steelseries_oled/models.py" = ["PLR2004"]  # Self-documenting byte thresholds in format_rate()

[tool.ruff.lint.isort]
known-first-party = ["steelseries_oled"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["hid.*", "psutil.*", "pynvml.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests requiring hardware",
]

[tool.coverage.run]
source = ["src/steelseries_oled"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
